refactor: `CafeInfo`로 넘어가는 props의 수 축소
